feat: add Pint configuration from pinkary.com
This commit adds the pint.json from pinkary.com to keep code style consistent. The changes from this config on all files are also in this commit. This also adds `* text=auto eol=lf` to .gitattributes to keep line endings consistent on windows with macos and unix to prevent unnecessary pint changes on windows.